Make the nelmio/api-doc-bundle recipe works out of the box
After applying this recipe, the route "/api/doc" was loaded by default, but it requires two component (`asset` and `twig`) that are usually not require for API only project.

The json route works of the box so it can be provided without additional action.

I faced it and it felt strange, and it seems that others have too : https://github.com/nelmio/NelmioApiDocBundle/issues/1220