Add Glide bundle configuration files (#1638)
* Add Glide bundle configuration files

Added a manifest.json and a dahromy_glide.yaml to configure the Glide bundle in Symfony. The manifest.json includes the bundle and its environment-specific parameters, while the dahromy_glide.yaml sets up source, cache locations and the glide signature key.

* Add 'gd' driver to dahromy_glide configuration

The gd library driver has been added to the dahromy_glide.yaml configuration file. This change enables image processing capabilities provided by the gd library in the dahromy/glide-symfony project.

* Add base_url to glide configuration

In the dahromy_glide.yaml configuration file, a new line specifying the base_url for image storage has been added. This amendment allows the glide library to know where to store and retrieve images.

* Add newline at the end of glide.yaml file

A newline has been added at the end of the dahromy_glide.yaml file. This is in accordance with standard coding practices and helps to prevent any potential parsing errors due to the absence of a newline.

* Remove 'gd' driver and base_url from glide config

The 'gd' driver and the 'base_url' configuration option have been removed from the Glide configuration. Other default configuration options can be added as needed. This change simplifies the default configuration.

* Move manifest and config files to 1.0 directory

All manifest and config files have been moved to a newly created 1.0 directory. This change is to improve file organization and version control. The structure change does not bring any functionality differences.

* Add PHP version conflict in manifest.json

The update adds a "conflict" field to the manifest.json which restricts the usage of the package with PHP versions less than 8.0. This ensures that the application is run on the supported PHP versions, preventing any compatibility issues.